move resolveBank into utils
